Implementando: Login Refatorado: Contrato para outra pasta, criando rotas para acesso.
