+++ package.json
@@ -23,6 +23,7 @@
     "@angular/platform-browser": "^9.1.0-next.4",
     "@angular/platform-browser-dynamic": "^9.1.0-next.4",
     "@angular/router": "^9.1.0-next.4",
+    "@auth0/angular-jwt": "^4.0.0",
     "angular2-text-mask": "^9.0.0",
     "classlist.js": "^1.1.20150312",
     "ngx-currency": "^2.2.2",
