+++ src/app/contrato/components/delete/delete.dialog.component.ts
@@ -1,17 +1,17 @@
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { Component, Inject } from '@angular/core';
-import { DataService } from '../../services/data.service';
+import { ContratoService } from '../../services/contrato.service';
 
 @Component({
-  selector: 'app-delete.dialog',
-  templateUrl: '../../dialogs/delete/delete.dialog.html',
-  styleUrls: ['../../dialogs/delete/delete.dialog.css']
+  selector: 'app-delete-contrato',
+  templateUrl: 'delete.dialog.html',
+  styleUrls: ['delete.dialog.css']
 })
 export class DeleteDialogComponent {
 
   constructor(public dialogRef: MatDialogRef<DeleteDialogComponent>,
               @Inject(MAT_DIALOG_DATA) public dataContrato: any,
-              public dataService: DataService) { }
+              public dataService: ContratoService) { }
 
   onNoClick(): void {
     this.dialogRef.close();
