+++ src/app/contrato/components/edit/edit.dialog.html
@@ -233,7 +233,7 @@
           </td>
           <td align="left" colspan="2">
             <!-- <div class="form"> -->
-              <app-file></app-file>
+              <app-file-contrato></app-file-contrato>
             <!-- </div> -->
           </td>
           <!-- Fim Grupo-->
