+++ src/app/contrato/components/file/file.dialog.component.scss
