+++ src/app/contrato/contrato.component.scss
@@ -1,0 +1,60 @@
+/* Structure */
+.container {
+	display: flex;
+	flex-direction: column;
+}
+
+/* Toolbar */
+.spacer {
+	flex: 1 1 auto;
+}
+
+/* Filter */
+.form {
+	min-height: 56px;
+	max-height: 56px;
+	display: flex;
+	align-items: center;
+	padding: 8px 24px 0;
+	font-size: 20px;
+	justify-content: space-between;
+	border-bottom: 1px solid transparent;
+}
+.mat-form-field {
+	font-size: 1vw; // 14px
+	flex-grow: 1;
+	margin-top: 8px;
+}
+
+/* Mat table */
+.no-results {
+	display: flex;
+	justify-content: center;
+	padding: 14px;
+	font-size: 14px;
+	font-style: italic;
+}
+.mat-cell {
+  font-size: 0.85vw; // Não existia
+	// &:nth-child(1) {
+	// 	flex: 0 0 2%;
+	// }
+	&:nth-child(1) {
+		flex: 0 0 22%;
+	}
+  &:nth-child(3) {
+		flex: 0 0 18%;
+	}
+}
+.mat-header-cell {
+  font-size: 0.9vw; // Não existia
+	// &:nth-child(1) {
+	// 	flex: 0 0 2%;
+	// }
+	&:nth-child(1) {
+		flex: 0 0 22%;
+	}
+  &:nth-child(3) {
+		flex: 0 0 18%;
+	}
+}
