+++ src/app/contrato/contrato.componentl.html
@@ -1,4 +1,3 @@
-<app-column-one></app-column-one>
 <mat-toolbar color="primary">
   <mat-toolbar-row>
     <span>Contratos</span>
