+++ src/app/models/auth.ts
@@ -1,5 +1,4 @@
-import { User } from './user';
 
 export class Auth {
-  user: User;
+  user: { token: string };
 }
