+++ src/app/shared/components/header/header.component.scss
@@ -1,0 +1,5 @@
+/* Toolbar */
+.spacer {
+  flex: 1 1 auto;
+  text-align: right;
+}
